requirements.txt
Django==4.2.7
django-cms==3.11.0
django-treebeard==4.7
django-classy-tags==4.1.0
django-sekizai==4.1.0
django-filer==3.1.0
easy-thumbnails==2.8.5
djangocms-admin-style==3.2.0
Pillow==10.1.0
django-crispy-forms==2.1
crispy-tailwind==0.5.0
django-ckeditor==6.7.0
django-meta==2.2.0
python-decouple==3.8
whitenoise==6.6.0
django-debug-toolbar==4.2.0
django-extensions==3.2.3
gunicorn==21.2.0